resnet-base/requirements.txt
scipy
numpy
h5py
tensorflow-gpu
foolbox==1.1.0
adversarial-vision-challenge
pillow
google-api-python-client
